Rename algorithm
